♻️ processDirecotory を directory.go として抽出
